backend/requirements.txt
google-generativeai
SpeechRecognition
pydub
python-dotenv
fastapi
uvicorn[standard]
python-multipart
pydantic
opencv-python
numpy
ultralytics
